Fixes ability to use Cloudformation intrinsic functions to specify SNS event sources.
It is sometimes necessary to use Cloudformation functions to reference SNS event sources created elsewhere, either within the current or other stacks. This feature was documented as working, but in practice an improper string check was stopping object values of the arn property. Additionally, a check to make sure usages of Fn::ImportValue did not include Fn::GetAtt or Ref was incomplete.

This fixes the logic to be in line with the documented feature, as well as adding a few relevant tests and updating the docs.